CHANGED PRICE FROM INT TO DOUBLE

+++ RazorPagesToMVCMigration/RazorPagesToMVCMigration/DAL/Repository/ProductRepository.cs
@@ -41,7 +41,7 @@ namespace RazorPagesToMVCMigration.DAL.Repository
                                 OEM = reader.GetString(0), // Column 1, VINNumber ...
                                 VINNumber = reader.GetString(1),
                                 Name = reader.GetString(2),
-                                Price = reader.GetInt32(3),
+                                Price = reader.GetDouble(3),
                                 DateAvailable = reader.GetDateTime(4),
                                 Notes = reader.GetString(5)
                             };
